feat: Fetch assignments for owner projects
This change introduces a new function `fetchAssignmentsForOwner` to retrieve all project assignments associated with projects created by a specific user.

This is utilized in the AdminDashboard to display more specific roles (e.g., "Line Producer") in the team list instead of a generic "User" when the user is a superuser and views projects they own.